Refactor: HouseDetailService Mock 테스트 분리 및 정리
